requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-decouple==2.1
django-social-share==1.3.2
djangorestframework==3.10.2
idna==2.8
oauthlib==3.1.0
Pillow==6.1.0
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
sqlparse==0.3.0
stripe==2.35.1
urllib3==1.25.3
